requirements.txt
beautifulsoup4
bs4
certifi
chardet
fake-useragent
idna
lxml
mypy
mypy-extensions
requests
soupsieve
typed-ast
typing-extensions
urllib3
python-telegram-bot
python-dotenv
